refactor dtype module structure
