+++ src/zarr/core/dtype/npy/common.py
@@ -1,13 +1,11 @@
 from __future__ import annotations
 
 import base64
-import struct
 import sys
 from collections.abc import Sequence
 from typing import (
     TYPE_CHECKING,
     Any,
-    Final,
     Literal,
     SupportsComplex,
     SupportsFloat,
@@ -15,17 +13,12 @@ from typing import (
     SupportsInt,
     TypeGuard,
     TypeVar,
+    get_args,
 )
 
 import numpy as np
 
-from zarr.core.dtype.common import (
-    ENDIANNESS_STR,
-    SPECIAL_FLOAT_STRINGS,
-    EndiannessStr,
-    JSONFloatV2,
-    JSONFloatV3,
-)
+from zarr.core.dtype.common import Endianness, JSONFloat
 
 if TYPE_CHECKING:
     from zarr.core.common import JSON, ZarrFormat
@@ -33,29 +26,9 @@ if TYPE_CHECKING:
 IntLike = SupportsInt | SupportsIndex | bytes | str
 FloatLike = SupportsIndex | SupportsFloat | bytes | str
 ComplexLike = SupportsFloat | SupportsIndex | SupportsComplex | bytes | str | None
-DateTimeUnit = Literal[
-    "Y", "M", "W", "D", "h", "m", "s", "ms", "us", "μs", "ns", "ps", "fs", "as", "generic"
-]
-DATETIME_UNIT: Final = (
-    "Y",
-    "M",
-    "W",
-    "D",
-    "h",
-    "m",
-    "s",
-    "ms",
-    "us",
-    "μs",
-    "ns",
-    "ps",
-    "fs",
-    "as",
-    "generic",
-)
-
-NumpyEndiannessStr = Literal[">", "<", "="]
-NUMPY_ENDIANNESS_STR: Final = ">", "<", "="
+DateUnit = Literal["Y", "M", "W", "D"]
+TimeUnit = Literal["h", "m", "s", "ms", "us", "μs", "ns", "ps", "fs", "as"]
+EndiannessNumpy = Literal[">", "<", "|", "="]
 
 TFloatDType_co = TypeVar(
     "TFloatDType_co",
@@ -72,18 +45,18 @@ TComplexDType_co = TypeVar(
 TComplexScalar_co = TypeVar("TComplexScalar_co", bound=np.complex64 | np.complex128, covariant=True)
 
 
-def endianness_from_numpy_str(endianness: NumpyEndiannessStr) -> EndiannessStr:
+def endianness_from_numpy_str(endianness: EndiannessNumpy) -> Endianness | None:
     """
     Convert a numpy endianness string literal to a human-readable literal value.
 
     Parameters
     ----------
-    endianness : Literal[">", "<", "="]
+    endianness : Literal[">", "<", "=", "|"]
         The numpy string representation of the endianness.
 
     Returns
     -------
-    Endianness
+    Endianness or None
         The human-readable representation of the endianness.
 
     Raises
@@ -99,21 +72,26 @@ def endianness_from_numpy_str(endianness: NumpyEndiannessStr) -> EndiannessStr:
             return "little"
         case ">":
             return "big"
-    raise ValueError(f"Invalid endianness: {endianness!r}. Expected one of {NUMPY_ENDIANNESS_STR}")
+        case "|":
+            # for dtypes without byte ordering semantics
+            return None
+    raise ValueError(
+        f"Invalid endianness: {endianness}. Expected one of {get_args(EndiannessNumpy)}"
+    )
 
 
-def endianness_to_numpy_str(endianness: EndiannessStr) -> NumpyEndiannessStr:
+def endianness_to_numpy_str(endianness: Endianness | None) -> EndiannessNumpy:
     """
     Convert an endianness literal to its numpy string representation.
 
     Parameters
     ----------
-    endianness : Endianness
+    endianness : Endianness or None
         The endianness to convert.
 
     Returns
     -------
-    Literal[">", "<"]
+    Literal[">", "<", "|"]
         The numpy string representation of the endianness.
 
     Raises
@@ -126,23 +104,14 @@ def endianness_to_numpy_str(endianness: EndiannessStr) -> NumpyEndiannessStr:
             return "<"
         case "big":
             return ">"
+        case None:
+            return "|"
     raise ValueError(
-        f"Invalid endianness: {endianness!r}. Expected one of {ENDIANNESS_STR} or None"
+        f"Invalid endianness: {endianness}. Expected one of {get_args(Endianness)} or None"
     )
 
 
-def get_endianness_from_numpy_dtype(dtype: np.dtype[np.generic]) -> EndiannessStr:
-    """
-    Gets the endianness from a numpy dtype that has an endianness. This function will
-    raise a ValueError if the numpy data type does not have a concrete endianness.
-    """
-    endianness = dtype.byteorder
-    if dtype.byteorder in NUMPY_ENDIANNESS_STR:
-        return endianness_from_numpy_str(endianness)  # type: ignore [arg-type]
-    raise ValueError(f"The dtype {dtype} has an unsupported endianness: {endianness}")
-
-
-def float_from_json_v2(data: JSONFloatV2) -> float:
+def float_from_json_v2(data: JSONFloat) -> float:
     """
     Convert a JSON float to a float (Zarr v2).
 
@@ -167,7 +136,7 @@ def float_from_json_v2(data: JSONFloatV2) -> float:
             return float(data)
 
 
-def float_from_json_v3(data: JSONFloatV3) -> float:
+def float_from_json_v3(data: JSONFloat) -> float:
     """
     Convert a JSON float to a float (v3).
 
@@ -180,41 +149,34 @@ def float_from_json_v3(data: JSONFloatV3) -> float:
     -------
     float
         The float value.
-
-    Notes
-    -----
-    Zarr V3 allows floats to be stored as hex strings. To quote the spec:
-       "...for float32, "NaN" is equivalent to "0x7fc00000".
-       This representation is the only way to specify a NaN value other than the specific NaN value
-       denoted by "NaN"."
     """
-
-    if isinstance(data, str):
-        if data in SPECIAL_FLOAT_STRINGS:
-            return float_from_json_v2(data)  # type: ignore[arg-type]
-        if not data.startswith("0x"):
-            msg = (
-                f"Invalid float value: {data!r}. Expected a string starting with the hex prefix"
-                " '0x', or one of 'NaN', 'Infinity', or '-Infinity'."
-            )
-            raise ValueError(msg)
-        if len(data[2:]) == 4:
-            dtype_code = ">e"
-        elif len(data[2:]) == 8:
-            dtype_code = ">f"
-        elif len(data[2:]) == 16:
-            dtype_code = ">d"
-        else:
-            msg = (
-                f"Invalid hexadecimal float value: {data!r}. "
-                "Expected the '0x' prefix to be followed by 4, 8, or 16 numeral characters"
-            )
-            raise ValueError(msg)
-        return float(struct.unpack(dtype_code, bytes.fromhex(data[2:]))[0])
+    # todo: support the v3-specific NaN handling
     return float_from_json_v2(data)
 
 
-def bytes_from_json(data: str, *, zarr_format: ZarrFormat) -> bytes:
+def float_from_json(data: JSONFloat, zarr_format: ZarrFormat) -> float:
+    """
+    Convert a JSON float to a float based on zarr format.
+
+    Parameters
+    ----------
+    data : JSONFloat
+        The JSON float to convert.
+    zarr_format : ZarrFormat
+        The zarr format version.
+
+    Returns
+    -------
+    float
+        The float value.
+    """
+    if zarr_format == 2:
+        return float_from_json_v2(data)
+    else:
+        return float_from_json_v3(data)
+
+
+def bytes_from_json(data: str, zarr_format: ZarrFormat) -> bytes:
     """
     Convert a JSON string to bytes
 
@@ -235,7 +197,7 @@ def bytes_from_json(data: str, *, zarr_format: ZarrFormat) -> bytes:
     # TODO: differentiate these as needed. This is a spec question.
     if zarr_format == 3:
         return base64.b64decode(data.encode("ascii"))
-    raise ValueError(f"Invalid zarr format: {zarr_format}. Expected 2 or 3.")  # pragma: no cover
+    raise ValueError(f"Invalid zarr format: {zarr_format}. Expected 2 or 3.")
 
 
 def bytes_to_json(data: bytes, zarr_format: ZarrFormat) -> str:
@@ -258,7 +220,7 @@ def bytes_to_json(data: bytes, zarr_format: ZarrFormat) -> str:
     return base64.b64encode(data).decode("ascii")
 
 
-def float_to_json_v2(data: float | np.floating[Any]) -> JSONFloatV2:
+def float_to_json_v2(data: float | np.floating[Any]) -> JSONFloat:
     """
     Convert a float to JSON (v2).
 
@@ -279,7 +241,7 @@ def float_to_json_v2(data: float | np.floating[Any]) -> JSONFloatV2:
     return float(data)
 
 
-def float_to_json_v3(data: float | np.floating[Any]) -> JSONFloatV3:
+def float_to_json_v3(data: float | np.floating[Any]) -> JSONFloat:
     """
     Convert a float to JSON (v3).
 
@@ -298,11 +260,9 @@ def float_to_json_v3(data: float | np.floating[Any]) -> JSONFloatV3:
     return float_to_json_v2(data)
 
 
-def complex_float_to_json_v3(
-    data: complex | np.complexfloating[Any, Any],
-) -> tuple[JSONFloatV3, JSONFloatV3]:
+def complex_to_json_v3(data: complex | np.complexfloating[Any, Any]) -> tuple[JSONFloat, JSONFloat]:
     """
-    Convert a complex number to JSON as defined by the Zarr V3 spec.
+    Convert a complex number to JSON (v3).
 
     Parameters
     ----------
@@ -317,15 +277,13 @@ def complex_float_to_json_v3(
     return float_to_json_v3(data.real), float_to_json_v3(data.imag)
 
 
-def complex_float_to_json_v2(
-    data: complex | np.complexfloating[Any, Any],
-) -> tuple[JSONFloatV2, JSONFloatV2]:
+def complex_to_json_v2(data: complex | np.complexfloating[Any, Any]) -> tuple[JSONFloat, JSONFloat]:
     """
-    Convert a complex number to JSON as defined by the Zarr V2 spec.
+    Convert a complex number to JSON (v2).
 
     Parameters
     ----------
-    data : complex | np.complexfloating
+    data : complex or np.complexfloating
         The complex value to convert.
 
     Returns
@@ -336,41 +294,55 @@ def complex_float_to_json_v2(
     return float_to_json_v2(data.real), float_to_json_v2(data.imag)
 
 
-def complex_float_from_json_v2(data: tuple[JSONFloatV2, JSONFloatV2]) -> complex:
+def complex_float_to_json(
+    data: complex | np.complexfloating[Any, Any], zarr_format: ZarrFormat
+) -> tuple[JSONFloat, JSONFloat]:
     """
-    Convert a JSON complex float to a complex number (v2).
+    Convert a complex number to JSON, parametrized by the zarr format version.
 
     Parameters
     ----------
-    data : tuple[JSONFloat, JSONFloat]
-        The JSON complex float to convert.
+    data : complex or np.complexfloating
+        The complex value to convert.
+    zarr_format : ZarrFormat
+        The zarr format version.
 
     Returns
     -------
-    np.complexfloating
-        The complex number.
+    tuple[JSONFloat, JSONFloat] or JSONFloat
+        The JSON representation of the complex number.
     """
-    return complex(float_from_json_v2(data[0]), float_from_json_v2(data[1]))
+    if zarr_format == 2:
+        return complex_to_json_v2(data)
+    else:
+        return complex_to_json_v3(data)
+    raise ValueError(f"Invalid zarr format: {zarr_format}. Expected 2 or 3.")
 
 
-def complex_float_from_json_v3(data: tuple[JSONFloatV3, JSONFloatV3]) -> complex:
+def float_to_json(data: float | np.floating[Any], zarr_format: ZarrFormat) -> JSONFloat:
     """
-    Convert a JSON complex float to a complex number (v3).
+    Convert a float to JSON, parametrized by the zarr format version.
 
     Parameters
     ----------
-    data : tuple[JSONFloat, JSONFloat]
-        The JSON complex float to convert.
+    data : float or np.floating
+        The float value to convert.
+    zarr_format : ZarrFormat
+        The zarr format version.
 
     Returns
     -------
-    np.complexfloating
-        The complex number.
+    JSONFloat
+        The JSON representation of the float.
     """
-    return complex(float_from_json_v3(data[0]), float_from_json_v3(data[1]))
+    if zarr_format == 2:
+        return float_to_json_v2(data)
+    else:
+        return float_to_json_v3(data)
+    raise ValueError(f"Invalid zarr format: {zarr_format}. Expected 2 or 3.")
 
 
-def check_json_float_v2(data: JSON) -> TypeGuard[JSONFloatV2]:
+def check_json_float_v2(data: JSON) -> TypeGuard[JSONFloat]:
     """
     Check if a JSON value represents a float (v2).
 
@@ -384,27 +356,12 @@ def check_json_float_v2(data: JSON) -> TypeGuard[JSONFloatV2]:
     Bool
         True if the data is a float, False otherwise.
     """
-    return data in ("NaN", "Infinity", "-Infinity") or isinstance(data, float | int)
+    if data == "NaN" or data == "Infinity" or data == "-Infinity":
+        return True
+    return isinstance(data, float | int)
 
 
-def check_json_float_v3(data: JSON) -> TypeGuard[JSONFloatV3]:
-    """
-    Check if a JSON value represents a float (v3).
-
-    Parameters
-    ----------
-    data : JSON
-        The JSON value to check.
-
-    Returns
-    -------
-    Bool
-        True if the data is a float, False otherwise.
-    """
-    return check_json_float_v2(data) or (isinstance(data, str) and data.startswith("0x"))
-
-
-def check_json_complex_float_v2(data: JSON) -> TypeGuard[tuple[JSONFloatV2, JSONFloatV2]]:
+def check_json_complex_float_v2(data: JSON) -> TypeGuard[tuple[JSONFloat, JSONFloat]]:
     """
     Check if a JSON value represents a complex float, as per the behavior of zarr-python 2.x
 
@@ -427,7 +384,25 @@ def check_json_complex_float_v2(data: JSON) -> TypeGuard[tuple[JSONFloatV2, JSON
     )
 
 
-def check_json_complex_float_v3(data: JSON) -> TypeGuard[tuple[JSONFloatV3, JSONFloatV3]]:
+def check_json_float_v3(data: JSON) -> TypeGuard[JSONFloat]:
+    """
+    Check if a JSON value represents a float (v3).
+
+    Parameters
+    ----------
+    data : JSON
+        The JSON value to check.
+
+    Returns
+    -------
+    Bool
+        True if the data is a float, False otherwise.
+    """
+    # TODO: handle the special JSON serialization of different NaN values
+    return check_json_float_v2(data)
+
+
+def check_json_complex_float_v3(data: JSON) -> TypeGuard[tuple[JSONFloat, JSONFloat]]:
     """
     Check if a JSON value represents a complex float, as per the zarr v3 spec
 
@@ -448,6 +423,51 @@ def check_json_complex_float_v3(data: JSON) -> TypeGuard[tuple[JSONFloatV3, JSON
         and check_json_float_v3(data[0])
         and check_json_float_v3(data[1])
     )
+
+
+def check_json_complex_float(
+    data: JSON, zarr_format: ZarrFormat
+) -> TypeGuard[tuple[JSONFloat, JSONFloat]]:
+    """
+    Check if a JSON value represents a complex float based on zarr format.
+
+    Parameters
+    ----------
+    data : JSON
+        The JSON value to check.
+    zarr_format : ZarrFormat
+        The zarr format version.
+
+    Returns
+    -------
+    Bool
+        True if the data represents a complex float, False otherwise.
+    """
+    if zarr_format == 2:
+        return check_json_complex_float_v2(data)
+    return check_json_complex_float_v3(data)
+
+
+def check_json_float(data: JSON, zarr_format: ZarrFormat) -> TypeGuard[float]:
+    """
+    Check if a JSON value represents a float based on zarr format.
+
+    Parameters
+    ----------
+    data : JSON
+        The JSON value to check.
+    zarr_format : ZarrFormat
+        The zarr format version.
+
+    Returns
+    -------
+    Bool
+        True if the data is a float, False otherwise.
+    """
+    if zarr_format == 2:
+        return check_json_float_v2(data)
+    else:
+        return check_json_float_v3(data)
 
 
 def check_json_int(data: JSON) -> TypeGuard[int]:
@@ -499,3 +519,60 @@ def check_json_bool(data: JSON) -> TypeGuard[bool]:
         True if the data is a boolean, False otherwise.
     """
     return isinstance(data, bool)
+
+
+def complex_float_from_json_v2(data: tuple[JSONFloat, JSONFloat]) -> complex:
+    """
+    Convert a JSON complex float to a complex number (v2).
+
+    Parameters
+    ----------
+    data : tuple[JSONFloat, JSONFloat]
+        The JSON complex float to convert.
+
+    Returns
+    -------
+    np.complexfloating
+        The complex number.
+    """
+    return complex(float_from_json_v2(data[0]), float_from_json_v2(data[1]))
+
+
+def complex_float_from_json_v3(data: tuple[JSONFloat, JSONFloat]) -> complex:
+    """
+    Convert a JSON complex float to a complex number (v3).
+
+    Parameters
+    ----------
+    data : tuple[JSONFloat, JSONFloat]
+        The JSON complex float to convert.
+
+    Returns
+    -------
+    np.complexfloating
+        The complex number.
+    """
+    return complex(float_from_json_v3(data[0]), float_from_json_v3(data[1]))
+
+
+def complex_float_from_json(data: tuple[JSONFloat, JSONFloat], zarr_format: ZarrFormat) -> complex:
+    """
+    Convert a JSON complex float to a complex number based on zarr format.
+
+    Parameters
+    ----------
+    data : tuple[JSONFloat, JSONFloat]
+        The JSON complex float to convert.
+    zarr_format : ZarrFormat
+        The zarr format version.
+
+    Returns
+    -------
+    np.complexfloating
+        The complex number.
+    """
+    if zarr_format == 2:
+        return complex_float_from_json_v2(data)
+    else:
+        return complex_float_from_json_v3(data)
+    raise ValueError(f"Invalid zarr format: {zarr_format}. Expected 2 or 3.")
